+++ packages/dumb/src/components/stats/time-in-range/cbg-percentage-bar/cbg-percentage-bar.hook.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2024, Diabeloop
+ * Copyright (c) 2022-2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -27,9 +27,9 @@
 
 import { useMemo } from 'react'
 import cbgTimeStatStyles from './cbg-percentage-bar.css'
-import stylesColors from '../common/cbg-colors.css'
-import { formatDuration } from '../../../utils/datetime/datetime.util'
-import { CBGStatType } from '../../../models/stats.model'
+import stylesColors from '../../common/cbg-colors.css'
+import { formatDuration } from '../../../../utils/datetime/datetime.util'
+import { CBGStatType } from '../../../../models/stats.model'
 
 export interface CBGPercentageBarHookProps {
   type: CBGStatType
@@ -37,6 +37,7 @@ export interface CBGPercentageBarHookProps {
   isDisabled: boolean
   total: number
   value: number
+  isReducedSize?: boolean
 }
 
 interface CBGPercentageBarHookReturn {
@@ -49,12 +50,13 @@ interface CBGPercentageBarHookReturn {
 }
 
 export const useCBGPercentageBar = (props: CBGPercentageBarHookProps): CBGPercentageBarHookReturn => {
-  const { type, id, isDisabled, total, value } = props
+  const { type, id, isDisabled, total, value, isReducedSize } = props
   const hasValues = total !== 0
   const percentage = hasValues ? Math.round(value / total * 100) : 0
   const rectangleBackgroundClass = isDisabled ? cbgTimeStatStyles['disabled-rectangle'] : stylesColors[`${id}-background`]
   const labelClass = isDisabled ? cbgTimeStatStyles['disabled-label'] : stylesColors[`${id}-color`]
-  const rectangleClasses = `${cbgTimeStatStyles.rectangle} ${rectangleBackgroundClass}`
+  const rectangleShapeClass = isReducedSize ? cbgTimeStatStyles['rectangle-reduced'] : cbgTimeStatStyles.rectangle
+  const rectangleClasses = `${rectangleShapeClass} ${rectangleBackgroundClass}`
   const durationBackgroundClasses =  isDisabled ? cbgTimeStatStyles['disabled-duration'] : stylesColors[`${id}-duration-background`]
   const barClasses = `${cbgTimeStatStyles['bar-value']} ${durationBackgroundClasses}`
   const percentageClasses = `${cbgTimeStatStyles['percentage-value']} ${labelClass}`
@@ -62,6 +64,7 @@ export const useCBGPercentageBar = (props: CBGPercentageBarHookProps): CBGPercen
   const barValue = useMemo(() => {
     switch (type) {
       case CBGStatType.TimeInRange:
+      case CBGStatType.TimeInTightRange:
         return formatDuration(value, true)
       case CBGStatType.ReadingsInRange:
         return (Math.round(value * 10) / 10).toString()